requirements.txt
fastapi[standard]
python-dotenv
jinja2
azure-functions